dataflow-types: split storage related modules into separate files
this commit is purely code movement
